Move test_dtype_device_mixin to lite (#14511)
